requirements.txt
django~=4.0.4
gunicorn
django-heroku
django-environ
whitenoise~=6.1.0
djangorestframework~=3.14.0
